feat: Add cursos property to user object
This commit adds a cursos property to the user object in the
loginUsuario controller. It retrieves the cursos from the database query
and assigns it to the user object.

refactor: Improve code clarity in getDuracionVideo function

This commit refactors the getDuracionVideo function in the tools file to
improve code clarity. It adds a null check for the video ID and only
calls the Youtube API if the ID is not null.