feat: Update app settings
This commit sets the "content-type" header to "application/json;
charset=utf-8" and enables "trust proxy" in the Express app.

refactor: Remove unused code in usuario.controller.ts

The unused code for creating tables "Usuario", "Usuario_Progreso", and
"Cliente_Usuario" has been removed.

fix: Set charset in login response

In the loginUsuario function, the charset is set to "utf-8" before
returning an error response for a non-verified user.

refactor: Update getCertificado function

The getCertificado function now accepts the "nombre" parameter in the
request body and uses it to draw the centered text on the certificate.